♻️ #205 - ActivitiesCalendar 컴포넌트 생성 및 적용
