+++ src/components/profile/index.ts
@@ -3,3 +3,4 @@ export * from './SelectProfileImage';
 export * from './ProfileImage';
 export * from './ActivitiesContainer';
 export * from './HeatmapDetail';
+export * from './ActivitiesCalendar';
